+++ changelog.md
@@ -18,6 +18,11 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 - add notifications for new location and update location in both form and modal
 - add confirmation modal for delete location
 - verification if an item open in the editor (event or news) has changes
+- "save" button is only available on change && all mandatory fields provided
+- events now have states "draft", "scheduled", "public" and "unpublished"
+- news now have states "draft", "scheduled", "public" and "unpublished"
+- items can be saved as draft (no check for mandatory fields or auto setting of dates)
+- "save as draft" is only available on changes
 
 ### Changed